Correcciones

+++ src/main/Prenda.java
@@ -3,9 +3,9 @@ package main;
 public class Prenda {
 	TipoPrenda tipoPrenda;
 	EstadoPrenda estado;
-	int precioBase;
+	double precioBase;
 	
-	public Prenda(TipoPrenda tipoPrenda, EstadoPrenda estado, int precioBase) {
+	public Prenda(TipoPrenda tipoPrenda, EstadoPrenda estado, double precioBase) {
 		this.tipoPrenda = tipoPrenda;
 		this.estado = estado;
 		this.precioBase = precioBase;
